+++ CHANGES.md
@@ -20,6 +20,7 @@ Features:
 
 Changes:
 - Retrieve lower court information in `bap9` and `bap10` #1567
+- Add lower court extraction to `federal_appellate` scrapers #1560
 
 Fixes:
 - fix `cafc` oral arguments scraper #1570